Tutorial: Create a structure with primitives
This tutorial is an exercise in which you can apply your understanding of GameObjects and transforms to build a structure out of 3D primitives. In this tutorial, you will:

Add primitives to an empty Scene.
Create an empty GameObject as a parent for other GameObjects.
Duplicate GameObjects.
Create parent-child relationships between GameObjects using the Hierarchy window.
Navigate around the Scene in the 3D space.
Transform the primitives to a given position, rotation, and scale.