Tutorial: Add components to 3D GameObjects
In this tutorial, you will see how components add functionality to GameObjects. The component you will see in action here, the Rigidbody component, lets a GameObject behave like a physical object instead of floating in space. In this tutorial, you will:
Position the Main Camera in order to achieve the desired framing of the scene.
Explore the relationship between GameObjects and components.
Identify specific components in the Inspector window.
Add physical properties to a 3D GameObject by applying the Rigidbody component and setting its properties.
Enter Play mode to preview a scene in the Game view.
Add a collider to a GameObject.
Experiment with 3D physics in Unity.