Tutorial: Manage GameObjects with prefabs - Part 1
Summary
In this tutorial, you’ll use prefabs to design, create, and organize many GameObjects in a scene. By the end of this tutorial, you will be able to:
Explain how to use prefabs in a scene.
Add a prefab to a project.
Edit a prefab in prefab mode.
Create a prefab variant.
Apply or revert changes to a prefab variant.
Nest prefabs.
Describe the process and outcomes for changing a nested prefab or prefab variant.