Tutorial: Get 3D assets
In this tutorial, we’ll explain sources of 3D assets: assets made with Digital Content Creation (DCC) tools or procured from the Unity Asset Store. We will walk through the steps for using the Asset Store to get assets you will need for the tutorial following this one.
By the end of this tutorial, you’ll be able to:
Explain the content creation tools that support real-time 3D production.
Navigate to the Asset Store in a web browser.
Search and filter a search in the Asset Store.
Import an asset from the Asset Store automatically via a Unity account.